refactor: close code-quality audit Group A (low-risk cleanups)
Eight scoped cleanups from docs/audit/code-quality.md, all verified
by `go test -race -tags integration` (76.1% coverage) and
`golangci-lint` (0 issues).

- oidc: replace init() gob.Register with sync.Once in oidc.New, so
  the side effect fires only when OIDC is configured.
- oauth: delete package-level tokenHMACKey + SetTokenHMACKey globals;
  thread hmacKey into oauth.Service. GenerateToken, ParseToken,
  hashToken, and parseTokenOrZero become methods on *Service.
- api: split toDocumentResponse's variadic-bool signature into
  toDocumentResponse + toDocumentResponseWithContent.
- api: fan out the 6 admin-dashboard count queries concurrently via
  errgroup.WithContext instead of 6 sequential round-trips.
- server: drop the unused Deps.IsSecure field (documented as
  reserved-for-future-use, never read).
- server: add a comment on the stacked rateLimitByIP middleware
  groups explaining the burst+sustained intent.
- oauth: extract UserCodePlaceholder ("XXXX-XXXX") constant shared
  by device verification template and tests.
- config: change Database.MaxOpenConns to int32 and drop both
  //nolint:gosec pragmas in foundation.go.

Incidental: observability integration test insertDocument now
sets status explicitly. The documents.status column default
'processing' predates migration 000014's CHECK constraint and
would fail the check on INSERT. Production code always sets
status via DocumentService.Create so the default is dead.

+++ internal/config/config_test.go
@@ -121,7 +121,7 @@ func TestLoad_Defaults(t *testing.T) {
 		{"Database.Host", cfg.Database.Host, "127.0.0.1"},
 		{"Database.Port", cfg.Database.Port, 5432},
 		{"Database.SSLMode", cfg.Database.SSLMode, "require"},
-		{"Database.MaxOpenConns", cfg.Database.MaxOpenConns, 25},
+		{"Database.MaxOpenConns", cfg.Database.MaxOpenConns, int32(25)},
 		{"Database.MaxIdleConns", cfg.Database.MaxIdleConns, 5},
 		{"Database.MaxLifetime", cfg.Database.MaxLifetime, 5 * time.Minute},
